mongoose: use websocket flag of connection struct
After the update of the mongoose webserver to version 7.8 new flags for
the connection struct were introduced. In the current version of
swupdate 'W' is set to the first position of the data array to mark
connections that use a websocket. For better readability the
is_websocket flag of the connection struct can be used.

Signed-off-by: Michael Bella <michael.bella@weidmueller.com>